requirements.txt
snowflake-snowpark-python==1.8.0
toml